🔨 Refactor/#151 - core에 있는 domain, repository을 event로 이동
