+++ src/main/java/com/daon/onjung/event/application/service/ProcessCompletedEventService.java
@@ -4,7 +4,7 @@ import com.daon.onjung.account.domain.Store;
 import com.daon.onjung.account.domain.User;
 import com.daon.onjung.account.domain.type.EBankName;
 import com.daon.onjung.account.repository.mysql.UserRepository;
-import com.daon.onjung.core.domain.service.ScheduledEventJobService;
+import com.daon.onjung.event.domain.service.ScheduledEventJobService;
 import com.daon.onjung.core.dto.CreateVirtualAccountResponseDto;
 import com.daon.onjung.core.exception.error.ErrorCode;
 import com.daon.onjung.core.exception.type.CommonException;
@@ -13,8 +13,8 @@ import com.daon.onjung.core.utility.DateTimeUtil;
 import com.daon.onjung.core.utility.FirebaseUtil;
 import com.daon.onjung.core.utility.RestClientUtil;
 import com.daon.onjung.event.application.usecase.ProcessCompletedEventUseCase;
-import com.daon.onjung.event.domain.Event;
-import com.daon.onjung.event.domain.Ticket;
+import com.daon.onjung.event.domain.mysql.Event;
+import com.daon.onjung.event.domain.mysql.Ticket;
 import com.daon.onjung.event.domain.service.EventService;
 import com.daon.onjung.event.domain.service.TicketService;
 import com.daon.onjung.event.repository.mysql.EventRepository;
