+++ src/main/java/com/daon/onjung/event/application/service/ReadTicketValidateService.java
@@ -6,10 +6,9 @@ import com.daon.onjung.core.utility.AESUtil;
 import com.daon.onjung.event.application.dto.request.ReadTicketValidateRequestDto;
 import com.daon.onjung.event.application.dto.response.ReadTicketValidateResponseDto;
 import com.daon.onjung.event.application.usecase.ReadTicketValidateUseCase;
-import com.daon.onjung.event.domain.Ticket;
+import com.daon.onjung.event.domain.mysql.Ticket;
 import com.daon.onjung.event.domain.service.TicketService;
 import com.daon.onjung.event.repository.mysql.TicketRepository;
-import com.daon.onjung.security.domain.mysql.Account;
 import com.daon.onjung.security.repository.mysql.AccountRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
