+++ src/main/java/com/daon/onjung/event/domain/service/TicketService.java
@@ -2,8 +2,8 @@ package com.daon.onjung.event.domain.service;
 
 import com.daon.onjung.account.domain.Store;
 import com.daon.onjung.account.domain.User;
-import com.daon.onjung.event.domain.Event;
-import com.daon.onjung.event.domain.Ticket;
+import com.daon.onjung.event.domain.mysql.Event;
+import com.daon.onjung.event.domain.mysql.Ticket;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
