+++ src/main/java/com/daon/onjung/security/application/service/SignUpOwnerByDefaultService.java
@@ -7,12 +7,12 @@ import com.daon.onjung.account.domain.service.StoreService;
 import com.daon.onjung.account.domain.type.EBankName;
 import com.daon.onjung.account.repository.mysql.OwnerRepository;
 import com.daon.onjung.account.repository.mysql.StoreRepository;
-import com.daon.onjung.core.domain.service.ScheduledEventJobService;
+import com.daon.onjung.event.domain.service.ScheduledEventJobService;
 import com.daon.onjung.core.dto.CreateVirtualAccountResponseDto;
 import com.daon.onjung.core.utility.BankUtil;
 import com.daon.onjung.core.utility.RestClientUtil;
 import com.daon.onjung.core.utility.S3Util;
-import com.daon.onjung.event.domain.Event;
+import com.daon.onjung.event.domain.mysql.Event;
 import com.daon.onjung.event.domain.service.EventService;
 import com.daon.onjung.event.repository.mysql.EventRepository;
 import com.daon.onjung.security.application.dto.request.SignUpOwnerByDefaultRequestDto;
